requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
semantic-kernel
openai
azure-identity

# If you see runtime errors about missing httpx, add:
# httpx
